# [Xamarin.Android.Build.Tasks] more ConvertResourcesCases improvements