@jonathanpeppers 2018-10-22
Build logs: [logs.zip](https://github.com/xamarin/xamarin-android/files/2503301/logs.zip)